added new undiscovered map icons. Moved map location objects to new file.
